feat: add error modal to favorite and complete toggle components
